requirements.txt
python-dateutil~=2.8.1
deprecation~=2.1.0
pytest
